improve how project builds
several changes:
* reorganized headers, there are now linux/ and bpf/ dirs,
  respectively for things from linux kernel and from libbpf
* removed dependency on vmlinux.h include, which was a massive
  file that included lots of things (types definitions,
  __sk_buff definition and more) but it was unclear what exactly
  was there
* introduced some basic Makefiles and building seems to work,
  though maybe it can be done better with cmake or simplified
  to one top-level Makefile - dunno, I'm not a Makefile god,
  but it builds, so I'm happy

this is necessary pre-work to move some of the code to lib headers,
to reduce bloat in bpf/nat64.c file, since it literally has all code
shoved in there and I don't like it. Makefiles should also come
in handy later when unit tests will be introduced.

Signed-off-by: Michal Siwinski <siwy@google.com>